requirements.txt
fastapi
uvicorn
xlsxwriter
requests
Pillow
pydantic
python-multipart